expAdd command refactor addExpForChannelMessage defaults removed, only manual now addUserExp module instead of helper
